move talent requests table component
